find and unpack misplaced glpyhs in TGA
If glpyhs in the atlas aren't placed in the grid quite right (like when
a new row is separated by more than one row of pink transparent pixels)
then I think the game will ignore them.

Previously, unpacking did the same. In this change, it's eager about
finding glpyhs and try to unpack those that are misplaced. However, if
that happens it still re-packes into a valid arrangement that won't
resemble the original TGA.

Also I didn't get around to testing it much so might be broken idk.
This behaviour probably should be optional. Use the last version if this
feature is bad.